taskにコメント一覧を表示、comment作った時のリダイレクトも修正

+++ db/schema.rb
@@ -26,7 +26,7 @@ ActiveRecord::Schema.define(version: 2021_03_15_012113) do
 
   create_table "comments", force: :cascade do |t|
     t.text "content", null: false
-    t.bigint "task_id", null: false
+    t.integer "task_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.index ["task_id"], name: "index_comments_on_task_id"
